REF: Extract experiments persistence helpers into _io
